Appear to have fixed the apparent memory leaks in the test cases. This was caused by the 'io' (TestIO) object being held onto by some global variables, and the 'io' object held some C-constructed packets.
So, those packets would hang around until the next test started.  HOWEVER, before the next test started, we looked
to make sure that none of the C-constructed packets were still around.

Solution:  Add a cleanio() member function to the TestIO class and call it before exiting the test.

This seems to have made the problem go away.  It is possible that some test needs it done, and I missed it, but I won't discover
it until things run in the "right" (wrong) order.

I did look for this case, and didn't see it.  We'll find out over time ;-)